Add deadlock detection mechanism on AbstractQueuedSynchronizer
The extension could not live in a separate class outside the AQS class
because it does direct usage of protected member variables such as tail
to iterate on the waiting queue of each lock. So instead of adding a new
class separated, its functions were merged into AQS and I've put them in
the end of the file for the sake of making obvious what is different from
the original code. AQS extends AOS, so I needed to change a AOS method to
non final in order to add a follow up event once the exclusive owner thread
is either set or unset to register and unregister owned locks.