Always throw exception on both threads involved in a deadlock
Extend the previous algorithm to keep a list of "tainted threads" on each lock.
A tainted thread is a thread that should throw a DeadlockException as soon as it
becomes the owner of that lock.

Basically when a thread detects a deadlock, it knows which locks that thread already have
that is desired by the conflicting thread, so for each of those locks, the conflicting thread
is added in the list of tainted threads. When the DeadlockException is raised, that lock will
be released and the conflicting thread will have the opportunity to acquire that lock. But as
soon as it does it, it will see itself in the list of tainted threads of that lock, thus raising
another DeadlockException.

The only issue of this approach is how to manage this list of thread references when both threads
raise exception simultaneously: this leaves no opportunity for the other thread to clean up those
references. As a damage control strategy, each thread that manages this list inside a lock will also
clean up every thread reference of threads that are no longer alive.