Only check for deadlock if conflictingThread is not null
If a thread is about to park and it discovers the owner of desired lock is null,
it means that it is being released or was just released. It is also safe to continue
with the parking procedure anyway as AbstractQueuedSynchronizer guarantees
that at this step a Thread can be successfully unparked later (if no deadlock happens).

We should not check for deadlock here. Reason: the desired lock is now free and soon
some other thread will acquire it. We can have two cases:
[1] Next thread to acquire it will be the current thread. No deadlock then.
[2] Next thread to acquire is another thread that was already on the queue.
If a thread B was already on the queue, then it means either:
[2.1] Thread B is about to park as well, so there's no TTTL deadlock (both threads
will be awaiting for the same lock)
[2.2] Thread B is fully already parked and before parking there was no TTTL deadlock.

If there is a deadlock happening with the current thread, then it should be caused
by one of the next threads acquiring the desired lock. Any of them would have to
first acquire the desired lock and then later try to acquire a lock owned by
current thread in order to succesfully cause a TTTL deadlock. So it will be detected
by the next thread checking for deadlock, thus there's no deadlock at this moment.